requirements.txt
# Python依赖包列表

# ========== 核心依赖 ==========

# Frida相关
frida-tools>=12.0.0
frida>=16.0.0

# HTTP请求
requests>=2.28.0

# 图片处理
Pillow>=9.0.0

# ========== 验证码识别（可选）==========

# OCR识别
ddddocr>=1.4.0
